tosca-os: Simplify discovery service implementation
The previous approach used a separate internal module for the `mdns-sd`
implementation, which added unnecessary complexity. Move the
implementation into a private module inside the `Service` file to keep
the discovery logic encapsulated and easier to maintain.
